# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Daily Creative API
  version: 0.1.0
  description: |
    Contract-first OpenAPI definition for Daily Creative.

    Version-one feature routes use the `/api/v1` prefix.
    Process health probes remain at the root (`/health/live`, `/health/ready`)
    and are not versioned under `/api/v1`.

    Authenticated feature endpoints use the `bearerAuth` security scheme
    (Descope-issued JWT). Health probes are unauthenticated.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/v1/me:
    get:
      operationId: getMe
      tags:
        - me
      summary: Current authenticated user
      description: |
        Returns the local application user for the authenticated Descope subject.
        On first successful authentication the backend provisions a local user
        with an incomplete profile and default preferences.
      security:
        - bearerAuth: []
      parameters:
        - name: creative_type
          in: query
          required: true
          description: Creative type context for preferences and profile stats.
          schema:
            $ref: "#/components/schemas/CreativeType"
          example: sketch
      responses:
        "200":
          description: Current user
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
              example:
                id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
                username: null
                display_name: New sketcher
                profile_completed: false
                status: incomplete
                avatar_url: null
                preferences:
                  notifications_enabled: false
                  notification_time_local: null
                  timezone: UTC
                  remember_timer_option: false
                  remembered_timer_mode: null
                  remembered_timer_seconds: null
                  appearance: system
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateMe
      tags:
        - me
      summary: Update current user profile
      description: |
        Updates the authenticated user's public profile fields.
        Setting a valid unique username and a non-empty display name completes
        the profile (`profile_completed` becomes true and status becomes active).
        Usernames are case-insensitive and must match `^[A-Za-z0-9_]{3,30}$`.
        Optionally attach a ready avatar upload via `avatar_upload_id`
        (purpose must be `avatar`). The upload is consumed on success.
      security:
        - bearerAuth: []
      parameters:
        - name: creative_type
          in: query
          required: true
          description: Creative type context for preferences and profile stats.
          schema:
            $ref: "#/components/schemas/CreativeType"
          example: sketch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
            example:
              username: sketchy_matt
              display_name: Matt
              bio: Sketching every day.
              avatar_upload_id: c3d4e5f6-a7b8-9012-cdef-123456789012
      responses:
        "200":
          description: Updated current user
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
              example:
                id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
                username: sketchy_matt
                display_name: Matt
                profile_completed: true
                status: active
                avatar_url: https://storage.example/avatar-display-signed
                preferences:
                  notifications_enabled: false
                  notification_time_local: null
                  timezone: UTC
                  remember_timer_option: false
                  remembered_timer_mode: null
                  remembered_timer_seconds: null
                  appearance: system
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteMe
      tags:
        - me
      summary: Request account deletion
      description: |
        Transitions the authenticated account to `pending_deletion`, immediately
        hides the public profile and Submissions, soft-deletes the user's
        Reflections, and removes the user's Likes. Access is disabled.
        Media and Descope identity cleanup run asynchronously via a scheduled
        finalize job. The request is idempotent: repeating while already
        pending returns the same confirmation. Optional `Idempotency-Key`
        may be supplied for client retries.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Account deletion accepted and pending finalization
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletionResponse"
              example:
                status: pending_deletion
                message: Your account deletion has been scheduled.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/me/preferences:
    get:
      operationId: getMyPreferences
      tags:
        - me
      summary: Current user preferences
      description: Returns server-backed preferences for the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - name: creative_type
          in: query
          required: true
          description: Creative type whose timer and session preferences are returned.
          schema:
            $ref: "#/components/schemas/CreativeType"
          example: sketch
      responses:
        "200":
          description: Preferences
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesSummary"
              example:
                notifications_enabled: true
                notification_time_local: "09:00:00"
                timezone: UTC
                remember_timer_option: true
                remembered_timer_mode: countdown
                remembered_timer_seconds: 300
                appearance: system
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateMyPreferences
      tags:
        - me
      summary: Update current user preferences
      description: |
        Partially updates preferences. Timer preferences must be consistent:
        `countdown` requires `remembered_timer_seconds` in {60, 180, 300, 600};
        `no_timer` requires `remembered_timer_seconds` to be null.
      security:
        - bearerAuth: []
      parameters:
        - name: creative_type
          in: query
          required: true
          description: Creative type whose timer and session preferences are updated.
          schema:
            $ref: "#/components/schemas/CreativeType"
          example: sketch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesUpdate"
            example:
              notifications_enabled: true
              notification_time_local: "09:00:00"
              remember_timer_option: true
              remembered_timer_mode: countdown
              remembered_timer_seconds: 300
              appearance: dark
      responses:
        "200":
          description: Updated preferences
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesSummary"
              example:
                notifications_enabled: true
                notification_time_local: "09:00:00"
                timezone: UTC
                remember_timer_option: true
                remembered_timer_mode: countdown
                remembered_timer_seconds: 300
                appearance: dark
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/feed/recent:
    get:
      operationId: getRecentFeed
      tags:
        - feed
      summary: Recent community feed
      description: |
        Returns a reverse-chronological page of published Submissions
        ordered by `published_at DESC, id DESC`. Soft-deleted, hidden,
        removed, and suspended/deleted-author content is excluded.
        Available to guests and authenticated users. When authenticated,
        `is_owner` and `viewer_has_liked` reflect the viewer.
        Pass an opaque `cursor` from a previous page to continue
        keyset pagination without duplicates or skips under concurrent
        inserts.
      parameters:
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous page, or omit for the first page.
          schema:
            type:
              - string
              - "null"
          example: null
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return (1–50). Defaults to 20.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          example: 20
        - name: creative_type
          in: query
          required: true
          description: Creative type of submissions to return.
          schema:
            $ref: "#/components/schemas/CreativeType"
          example: sketch
      responses:
        "200":
          description: Feed page
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentFeed"
              examples:
                empty:
                  summary: Empty community feed
                  value:
                    items: []
                    next_cursor: null
                populated:
                  summary: Populated reverse-chronological page
                  value:
                    items:
                      - id: d4e5f6a7-b8c9-0123-def0-234567890123
                        image_url: https://storage.example/display-signed
                        thumbnail_url: https://storage.example/thumb-signed
                        user:
                          id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
                          username: sketchy_matt
                          display_name: Matt
                          avatar_url: null
                        prompt:
                          id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                          prompt_date: "2026-07-18"
                          word_1: Chocolate
                          word_2: Coffee
                          word_3: Banana
                        timer_mode: countdown
                        timer_seconds: 300
                        caption_preview: A quiet coffee and banana sketch.
                        like_count: 0
                        reflection_count: 0
                        viewer_has_liked: false
                        is_owner: false
                        published_at: "2026-07-18T20:12:00Z"
                    next_cursor: null
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Descope-issued access/session JWT sent as Authorization Bearer token.
  headers:
    RequestId:
      description: |
        Opaque request correlation identifier. Clients may supply
        `X-Request-ID`; otherwise the server generates one. The same
        value is returned on every response and embedded in error bodies
        as `error.request_id`.
      required: true
      schema:
        type: string
        format: uuid
      example: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-generated key that makes a write operation safely retryable.
        Replaying the same key with the same request body returns the original
        response. Replaying with a different body returns `idempotency_key_conflict`.
      schema:
        type: string
        minLength: 1
        maxLength: 255
      example: sketch-session-create-7f3a9c2e
  responses:
    Unauthenticated:
      description: Missing or invalid bearer token
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthenticated
              message: Authentication is required.
              details: {}
              request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    Forbidden:
      description: Authenticated but not permitted (suspended or unavailable account)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            account_suspended:
              summary: Suspended account
              value:
                error:
                  code: account_suspended
                  message: This account has been suspended.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            account_unavailable:
              summary: Deleted or pending deletion
              value:
                error:
                  code: account_unavailable
                  message: This account is no longer available.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            profile_incomplete:
              summary: Profile incomplete
              value:
                error:
                  code: profile_incomplete
                  message: Complete your profile before publishing.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_forbidden:
              summary: Not the Reflection author
              value:
                error:
                  code: reflection_forbidden
                  message: You can only delete your own reflection.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    Conflict:
      description: Request conflicts with existing state
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            username_taken:
              summary: Username already taken
              value:
                error:
                  code: username_taken
                  message: That username is already taken.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            idempotency_key_conflict:
              summary: Idempotency key reused with a different body
              value:
                error:
                  code: idempotency_key_conflict
                  message: This idempotency key was already used with a different request.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            upload_already_consumed:
              summary: Upload already consumed
              value:
                error:
                  code: upload_already_consumed
                  message: This upload has already been used.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            session_already_submitted:
              summary: Sketch Session already has a Submission
              value:
                error:
                  code: session_already_submitted
                  message: This sketch session already has a published submission.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    NotFound:
      description: Requested resource was not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            user_not_found:
              summary: Public profile not found
              value:
                error:
                  code: user_not_found
                  message: The requested profile could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            prompt_not_found:
              summary: Daily Prompt not found
              value:
                error:
                  code: prompt_not_found
                  message: Today's prompt is not available yet.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            session_not_found:
              summary: Sketch Session not found
              value:
                error:
                  code: session_not_found
                  message: The requested sketch session could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            upload_not_found:
              summary: Upload not found
              value:
                error:
                  code: upload_not_found
                  message: The requested upload could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            submission_not_found:
              summary: Submission not found
              value:
                error:
                  code: submission_not_found
                  message: The requested sketch could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_not_found:
              summary: Reflection not found
              value:
                error:
                  code: reflection_not_found
                  message: The requested reflection could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    ValidationError:
      description: Request body failed validation or domain rules
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            username_invalid:
              summary: Invalid username format
              value:
                error:
                  code: username_invalid
                  message: Usernames must be 3–30 characters and use only letters, numbers, and underscores.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            username_reserved:
              summary: Reserved username
              value:
                error:
                  code: username_reserved
                  message: That username is reserved.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_timer_preference:
              summary: Invalid timer preference combination
              value:
                error:
                  code: invalid_timer_preference
                  message: Timer preference mode and seconds are inconsistent.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_timer_selection:
              summary: Invalid Sketch Session timer selection
              value:
                error:
                  code: invalid_timer_selection
                  message: Timer mode and selected seconds are inconsistent.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_session_transition:
              summary: Invalid Sketch Session lifecycle transition
              value:
                error:
                  code: invalid_session_transition
                  message: That lifecycle event is not valid for the current session state.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            upload_not_ready:
              summary: Upload not ready for consumption
              value:
                error:
                  code: upload_not_ready
                  message: This upload is not ready to publish yet.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_image:
              summary: Image could not be decoded or validated
              value:
                error:
                  code: invalid_image
                  message: The uploaded file is not a valid image.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            image_too_large:
              summary: Image exceeds size limit
              value:
                error:
                  code: image_too_large
                  message: That image is too large to upload.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            unsupported_media_type:
              summary: Unsupported image content type
              value:
                error:
                  code: unsupported_media_type
                  message: That image type is not supported.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            object_missing:
              summary: Storage object missing after direct upload
              value:
                error:
                  code: object_missing
                  message: The uploaded image could not be found. Please try again.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            validation_error:
              summary: Schema validation failure
              value:
                error:
                  code: validation_error
                  message: The request could not be validated.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_body_empty:
              summary: Empty or whitespace-only Reflection
              value:
                error:
                  code: validation_error
                  message: The request could not be validated.
                  details:
                    body: empty
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_body_too_long:
              summary: Reflection exceeds character limit
              value:
                error:
                  code: validation_error
                  message: The request could not be validated.
                  details:
                    body: too_long
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    InternalError:
      description: Unexpected server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal_error
              message: An unexpected error occurred.
              details: {}
              request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
...
  schemas:
    UUID:
      type: string
      format: uuid
      description: Canonical UUID string identifier.
      example: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
...
    UserStatus:
      type: string
      description: Local application account lifecycle status.
      enum:
        - incomplete
        - active
        - suspended
        - pending_deletion
        - deleted
...
    PreferencesSummary:
      type: object
      description: |
        Preference summary returned with the current user and preferences endpoints.
        Values are persisted in `user_preferences`.
      required:
        - notifications_enabled
        - notification_time_local
        - timezone
        - remember_timer_option
        - remembered_timer_mode
        - remembered_timer_seconds
        - appearance
      properties:
        notifications_enabled:
          type: boolean
          example: false
        notification_time_local:
          type:
            - string
            - "null"
          description: Local reminder time as HH:MM:SS, or null when unset.
          example: null
        timezone:
          type: string
          example: UTC
        remember_timer_option:
          type: boolean
          example: false
        remembered_timer_mode:
          oneOf:
            - $ref: "#/components/schemas/TimerMode"
            - type: "null"
          example: null
        remembered_timer_seconds:
          type:
            - integer
            - "null"
          example: null
        appearance:
          $ref: "#/components/schemas/AppearancePreference"
    PreferencesUpdate:
      type: object
      description: Partial preferences update. Omitted fields are left unchanged.
      additionalProperties: false
      properties:
        notifications_enabled:
          type: boolean
        notification_time_local:
          type:
            - string
            - "null"
          description: Local reminder time as HH:MM:SS, or null to clear.
          example: "09:00:00"
        timezone:
          type: string
          example: UTC
        remember_timer_option:
          type: boolean
        remembered_timer_mode:
          oneOf:
            - $ref: "#/components/schemas/TimerMode"
            - type: "null"
        remembered_timer_seconds:
          type:
            - integer
            - "null"
          description: Allowed countdown values are 60, 180, 300, or 600.
        appearance:
          $ref: "#/components/schemas/AppearancePreference"
    UpdateMeRequest:
      type: object
      description: |
        Partial profile update. Omitted fields are left unchanged.
        When `avatar_upload_id` is supplied, the upload must be owned by the
        caller, have `purpose: avatar`, and be in `ready` status (or already
        the caller's current avatar for idempotent retries). It is consumed
        on success. Errors include `upload_not_found`, `upload_not_ready`,
        `upload_already_consumed`, and `avatar_upload_invalid`.
      additionalProperties: false
      properties:
        username:
          type: string
          description: |
            Public username. Must match `^[A-Za-z0-9_]{3,30}$`.
            Uniqueness is case-insensitive.
          minLength: 3
          maxLength: 30
          example: sketchy_matt
        display_name:
          type: string
          minLength: 1
          maxLength: 120
          example: Matt
        bio:
          type:
            - string
            - "null"
          description: Optional short bio. Null clears an existing bio.
          maxLength: 280
          example: Sketching every day.
        avatar_upload_id:
          description: Ready avatar upload to attach as the current avatar.
          allOf:
            - $ref: "#/components/schemas/UUID"
          example: c3d4e5f6-a7b8-9012-cdef-123456789012
    CurrentUser:
      type: object
      required:
        - id
        - username
        - display_name
        - profile_completed
        - status
        - avatar_url
        - preferences
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        username:
          type:
            - string
            - "null"
          description: Public username; null until profile completion.
          example: null
        display_name:
          type: string
          example: New sketcher
        profile_completed:
          type: boolean
          description: True when profile_completed_at is set.
          example: false
        status:
          $ref: "#/components/schemas/UserStatus"
        avatar_url:
          type:
            - string
            - "null"
          description: Signed avatar display URL when an avatar is set; otherwise null.
          format: uri
          example: null
        preferences:
          $ref: "#/components/schemas/PreferencesSummary"
...
    CreativeType:
      type: string
      description: The type of creative activity.
      enum:
        - sketch
        - story
...
    AccountDeletionResponse:
      type: object
      description: Confirmation that account deletion has been accepted.
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum:
            - pending_deletion
          example: pending_deletion
        message:
          type: string
          example: Your account deletion has been scheduled.
...